+++ term-project/musiviz/server.bat
@@ -1,0 +1,1 @@
+py -m http.server
